+++ sc/social/like/plugins/telegram/plugin.pt
@@ -1,0 +1,5 @@
+<tal:telegram i18n:domain="sc.social.like">
+  <a tal:attributes="href view/telegramurl;
+                     class view/klass"
+      i18n:translate="">Share</a>
+</tal:telegram>
